Update babel-preset to include 'platform' option
Passing the 'platform' to the babel-preset will ensure the correct
plugins are used. By default, the platform is 'web'.

Also updates stylex to 0.9.3
