+++ apps/examples/babel.config.js
@@ -26,12 +26,11 @@ module.exports = function (api) {
   const platform = api.caller(getPlatform);
   const dev = api.caller(getIsDev);
 
-  const presets = ['babel-preset-expo'];
   const plugins = [];
-
-  if (platform === 'web') {
-    presets.push([reactStrictPreset, { debug: true, dev, rootDir: __dirname }]);
-  }
+  const presets = [
+    'babel-preset-expo',
+    [reactStrictPreset, { debug: true, dev, platform }]
+  ];
 
   return {
     plugins,
